Nav Search Input Control Overlay
-Nav Bar wider than usual Nav Bar
-Improve layout of search control to optimise for smaller

#2546
