--- waltz-ng/client/navbar/components/navbar-search-form.html
@@ -1,24 +1,0 @@
-<ul class="nav navbar-nav navbar-form navbar-right">
-    <li>
-        <form role="search">
-            <div class="form-group">
-                <div class="input-group"
-                     style="display: table;">
-                    <input type="text"
-                           class="form-control"
-                           ng-model="$ctrl.query"
-                           ng-model-options="{ debounce: 250 }"
-                           ng-change="$ctrl.doSearch()"
-                           autocomplete="off"
-                           autofocus="autofocus"
-                           placeholder="Search">
-                </div>
-            </div>
-        </form>
-    </li>
-    <waltz-nav-search-results results="$ctrl.searchResults"
-                              query="{{ $ctrl.query }}"
-                              on-dismiss="$ctrl.dismissResults"
-                              ng-if="$ctrl.showSearch()">
-    </waltz-nav-search-results>
-</ul>
