+++ waltz-ng/client/navbar/directives/navbar.html
@@ -31,7 +31,7 @@
         <waltz-navbar-recently-viewed>
         </waltz-navbar-recently-viewed>
 
-        <waltz-navbar-search-form>
-        </waltz-navbar-search-form>
+        <waltz-navbar-search>
+        </waltz-navbar-search>
     </div>
 </nav>
