+++ waltz-ng/client/navbar/styles.scss
@@ -1,1 +1,1 @@
-@import "./components/nav-search-results";
+@import "components/nav-search-overlay";
